Compose API and web handlers in server
